Fixes bugs and everything is working. Adds assembly loading from assemblies folder.
