--- Data/Assemblies.cfg
@@ -1,7 +1,0 @@
-System.dll
-System.Core.dll
-System.Data.dll
-System.Drawing.dll
-System.Web.dll
-System.Windows.Forms.dll
-System.Xml.dll
