+++ Scripts/Misc/DataPath.cs
@@ -10,8 +10,8 @@ namespace Server.Misc
      * Example:
      *  private static string CustomPath = @"C:\Program Files\Ultima Online";
      */
-    //private static string CustomPath = @"C:\Ultima Online Classic";
-    private static string CustomPath = @"/Users/kamronbatman/UOC";
+    private static string CustomPath = @"C:\Ultima Online Classic";
+    // private static string CustomPath = @"/Users/kamronbatman/UOC";
 
     /* The following is a list of files which a required for proper execution:
      *
