+++ Server/Network/Listener.cs
@@ -61,9 +61,7 @@ namespace Server.Network
         Socket s;
         try
         {
-          Console.WriteLine("Listener: Waiting for Socket...");
           s = await m_Socket.AcceptAsync().ConfigureAwait(false);
-          Console.WriteLine("Listener: Socket Connection Detected!");
         }
         catch (SocketException ex)
         {
