+++ Server/Network/SocketExtensions.cs
@@ -1,3 +1,23 @@
+/***************************************************************************
+ *                            SocketExtensions.cs
+ *                            -------------------
+ *   begin                : August 2, 2019
+ *   copyright            : (C) The ModernUO Team
+ *   email                : hi@modernuo.com
+ *
+ *   $Id$
+ *
+ ***************************************************************************/
+
+/***************************************************************************
+ *
+ *   This program is free software; you can redistribute it and/or modify
+ *   it under the terms of the GNU General Public License as published by
+ *   the Free Software Foundation; either version 2 of the License, or
+ *   (at your option) any later version.
+ *
+ ***************************************************************************/
+
 using System;
 using System.Buffers;
 using System.Net.Sockets;
